refactor(mocks): consolidate nest-commander mock and eliminate double app creation
Centralize nest-commander mock in test/__mocks__/nest-commander.ts
to replace inline jest.mock blocks in main.spec.ts and cli.module.spec.ts.

Replace NestFactory.createApplicationContext + CommandFactory.run pattern
with single app lifecycle via createWithoutRunning() and runApplication().
Move dependency check logic inline from deleted setup.ts.
