+++ test/__mocks__/nest-commander.ts
@@ -1,0 +1,5 @@
+export const CommandFactory = {
+  createWithoutRunning: jest.fn(),
+  run: jest.fn().mockResolvedValue(undefined),
+  runApplication: jest.fn().mockResolvedValue(undefined)
+}
